API references

LEDStrip.SPIStrip — Type

struct SPIStrip
    spi::Int                        # SPI number 1 or 2
    pixel_count::Int                # number of pixels in RGB Strip, equals to buffer size
    freq::Int                       # frequency in Hz for sending messages to WS281x chip
    color_indexes::Vector{Int}      # color indexes in RGB notation
    buffer::AbstractArray{UInt32}   # array storing states of pixels
end

Object storing settings and states of addressable LED Strip.

SPIStrip includes the intermediate buffer which is a variable storing pixel colors in RGB.

There are two ways working with SPIStrip:

Sending colors directly to Strip using the methods like send_bytes, send_colors.

Updating buffer values and make changes visible after that using methods: set_pixels!, clean_pixels!, show_pixels, etc.

LEDStrip.SPIStrip — Method

SPIStrip(
    spi::Int;
    pixel_count::Int = 10,
    freq::Int = 800_000,
    color_type::String = "grb"
)

Creates object representing LED Strip.

Arguments

spi: number of SPI used: 1 or 2.

1 value corresponds to main SPI device. Data pin of LED Strip should be connected to MOSI (GPIO10), MISO (GPIO9), CLK (GPIO11), CC0 (GPIO8), CC1 (GPIO7) are not used for LED Strip.

2 value corresponds to alternative SPI device. Data pin of LED Strip should be connected to MOSI (GPIO20), MISO (GPI19), CLK (GPIO21), CC0 (GPI18), CC1 (GPI17), CC2 (GPIO16) are not used for LED Strip.

pixel_count: number of physical pixels (RGB LED items)

This value equals to buffer size.

freq: signal rate of WS281x chip. Default 800000 Hz.

Usually RGB LED Strip support 800 000 Hz messages (sometimes 400 000 Hz).

color_type: RGB sequence may vary on some strips. Default is grb

Valid values are "rgb", "rbg", "grb", "gbr", "bgr", "brg".

LEDStrip.clean_pixels! — Method

clean_pixels!(strip::AbstractStrip)

This method removes colors from whole Strip buffer or in other words set them as black. This is the same as setting all buffer values as 0x000000. To make it visible you must use show_pixels after.

Arguments

strip: SPIStrip object

LEDStrip.hide_pixels — Method

hide_pixels(strip::AbstractStrip)

Clean physical buffer colors i.e. set them as black. This method does not change the buffer state. You can hide and show colors but buffer content will be the same.

Arguments

strip: SPIStrip object

LEDStrip.send_bytes — Method

send_bytes(
    strip::SPIStrip,
    byte_array::AbstractArray{UInt8}
)

Sending sequence of bytes to LED Strip. Each byte regulates brightness of one LED. LED strip consists of pixels under control of the WS281x chip: tree LEDs per pixel. To send the color "pixel-wisely" use send_colors method.

The method does not update the internal buffer of Strip object.

Arguments

strip: SPIStrip object

byte_array: Array of bytes UInt8 to send. First byte will be send to first LED (green), the second one will be send to second LED (red), etc. The byte's value regulates the brightness of corresponding LED. Three bytes together regulates the color of pixel.

LEDStrip.send_colors — Method

send_colors(
    strip::AbstractStrip,
    color_array::AbstractArray{UInt32}
)

Send message to set Strip pixel colors in RGB notation.

The method does not update the internal buffer of Strip object.

Example

send_colors(strip, [0x000000, 0xffffff, 0xff0000])`

The first pixel is red, second is white, third is red the other pixels will not be updated.

Arguments

strip: SPIStrip object

color_array: Each color is an unsigned 32-bit value (UInt32) where the lower 24 bits define the red, green, blue data (each being 8 bits long).

LEDStrip.set_pixel! — Method

set_pixel!(
    strip::AbstractStrip,
    num::Int,
    color::UInt32
)

This method updates the color of particular pixel in Strip's buffer. To make it visible you must use show_pixels after updating.

Example

set_pixel(strip, 10, 0x00ff00)
show_pixels(strip);

Set pixel #10 as green and make it visible.

Arguments

strip: SPIStrip object

num: number of pixel in strip starting from 1.

color: UInt32 value representing color and brightness.

LEDStrip.set_pixels! — Method

set_pixels!(
    strip::AbstractStrip,
    arr::AbstractArray{UInt32};
    replicate::Bool = false
)

This method updates whole pixel buffer in Strip. To make it visible you must use show_pixels after updating.

Arguments

strip: SPIStrip object

color_arr: UInt32 vector representing color and brightness for all pixels in Strip. If this vector is longer than the buffer length the rest values will be ignored. If color_arr is shorter the result will depend on replicate argument.

replicate: If true the shorter color_arr will be replicated to the end of buffer.

LEDStrip.shift_backward! — Method

shift_backward!(
    strip::AbstractStrip;
    circular::Bool = false
)

Shifts Strip buffer values in negative direction, i.e. pixel #2 will be moved to pixel #1, #3 => #2, #4 => #3 etc. This can be used to create animation.

Arguments

strip: SPIStrip object

circular: If true the last pixel will be set as first. If false the last value will be 0x000000 (black).

LEDStrip.shift_forward! — Method

shift_forward!(
    strip::AbstractStrip;
    circular::Bool = false
)

Shifts Strip buffer values forward, i.e. pixel #1 will be moved to pixel #2, #2 => #3, #3 => #4 etc. This can be used to create animation.

Arguments

strip: SPIStrip object

circular: If true the #1 pixel will be set as the last one. If false the first value will be 0x000000 (black).

LEDStrip.show_pixels — Method

show_pixels(strip::AbstractStrip)

Display buffer content in pixels. Updating colors by set_pixels!, clean_pixels!, etc. changes only buffer state. This method make it visible in physical pixels.

Arguments

strip: SPIStrip object